First attempt at recording for testing.
This currently works as is, by setting the env variables:
* COOKBOOK_RECORDING_ENABLED -> to record the session
* COOBKOOK_REPLAYING_ENABLED -> to replay a session
* COOKBOOK_RECORDING_FILE -> where to store/read the recording

More details of usage in the README.md

It includes a pytest fixture to run a cookbook with the given record
file.

It will raise if the record file is not fully replayed too, as a
sanity check.

Also re-shuffled a bit the test files to match the modules and
separated functional from unit tests.

Some things that will come next:
* Improve default file name.
* Make it work also for API calls (alertmanager, future openstack)

Signed-off-by: David Caro <dcaro@wikimedia.org>
Change-Id: If3b6b98b11cff147a6ee9ad13efd01deea6d5e23