wmcs_libs: inventory: add toolforge k8s clusters
Adds inventory data for Toolforge Kubernetes clusters, so that we can
update cookbooks to take a single 'cluster' param and not two or three
for the cluster settings.

Note this assumes that there is at most one cluster per project. I think
that's fine for now, and if that ever changes we need to update quite a
few different places not including this one.

Change-Id: I20d51333c4cb2fd0fee4d0e3115ae75dccc23f03