Support connecting to wm-bot through proxy
When running in cloudcuminXXXX, connections to wm-bot.wm-bot.wmcloud.org
(used for SAL logging) must go through the webproxy, because
cloudcumin hosts are not allowed to open direct TCP connections to the
internet, or to Cloud VPS.

This patch uses the existing spicerack.http_proxy config option, that
can be set in /etc/spicerack/config.yaml, to determine if a proxy is
required for outbound connections.

Spicerack manages the proxy automatically for calls using the "requests"
library, but the SALLogger class uses a low-level "socket" instead, so
we have to manually add the call to the proxy.

Note: I removed the code that tries multiple addresses to reach the
wm-bot host, as it seems to work fine without it, but it can be
reinstated if necessary.

Bug: T325756
Change-Id: I4e575a6361d50509fb1b40aff923d34e4b15c955