Add project name to Spicerack's sal_logger
This adds the project name as a prefix to the SAL log messages generated
by Spicerack, so that they can be archived under the correct project.

This has not been a problem so far because Spicerack's sal_logger is
disabled when running Spicerack from a laptop, but it is enabled when we
run Spicerack from the new cloudcumin hosts (T319401).

If this works, we could use Spicerack's sal_logger in place of our
custom SALLogger implemented in wmcs_libs/common.py

Bug: T341793
Change-Id: I1813b07f493c014b611df4da5fb398e0c76bd901