toolforge: k8s: use verbatism reasons for puppet disables
This turns on the verbatism_reason flag for the Kubernetes upgrade
cookbooks to allow a different user to enable Puppet compared to the one
who disabled it.

This is not ideal as I'd like to include the username in the reason, but
the spicerack API does not seem to expose a way to get the reason and we
don't want to override unrelated Puppet disables.

Bug: T343869
Change-Id: Ib8925f34b810ed721b797c440ab9af1bef036b7d