wmcs_libs: Fix spicerack.sal_logger on local setups
Configures the Spicerack native sal_logger to relay messages via wm-bot
if a logmsgbot integration has not been configured. This is meant to
ensure that cloudcumin and local environments behave in a similar way
(that is, emit the START/END messages) until Spicerack has native
support for wm-bot style messages.

Change-Id: I2d80fdfe9d4b5f8b1cc63159282506996a3c1836