ceph: add drain/undrain host and rack cookbooks
This might not be needed once we fix all the hard drives, but it's
good to have in any case for safe/gentle drain/undrain.

Had to remove the osd functional tests as I changed the commands it
runs, will have to redo next time I can run those cookbooks on codfw1
or similar.

Bug: T329709
Signed-off-by: David Caro <dcaro@wikimedia.org>
Change-Id: Ieb24e18164859ed1920cc8948a51a8d441f93a56