set_maintenance: do not downtime host
The set_maintenance cookbook does not need to downtime the host in
icinga or alertmanager, because no alert is raised when a host is
assigned to the "maintenance" aggregate (e.g. when draining a host).

Removing the downtime also means we can run this cookbook from
cloudcumin, where downtiming is currently broken (T347490)

Change-Id: Ib812e8761e707fd49ecaf2c9cc8b0783fdb772a6