ceph: drain and undrain in chunks
This makes adding and removing osds work in batches, to avoid
saturating the network.

Sets `noin` also to avoid the mons calculating shuffles before we
start adding the osds.

Sets the default to wait for rebalance too.

It also unifies the parameters for both cookbooks to be more similar.

Had to refactor the cookbooks to avoid linting issues.

Bug: T329709
Signed-off-by: David Caro <dcaro@wikimedia.org>
Change-Id: Ia0b4c93ac5ccc5df18dae7d8d5c48301701641ea
Signed-off-by: David Caro <dcaro@wikimedia.org>