wmcs_libs.common: add run_script
Sometimes we want to run more than one command at a time, this allows
running an inline "script", also using correctly the sudo parameters.

Adding the user parameter to both run_one_raw and run_script allows us
to run as a specific user (kinda ugly how it's implemented... but it
comes down to how spicerack handles commands, just concatenates them
as strings).

See the next patch of the stack for the rationale and usage.

Signed-off-by: David Caro <dcaro@wikimedia.org>
Change-Id: I4bf6cfc8403cda9962d82f30973d53edb087e6b0