toolforge.component.deploy: fix when there's many forks with same name
If there's other repos with the same name (ex. forks), when getting
the package build jobs we were just getting the first matching repo,
that might not be the one in the toolforge org.

This adds that filter so we check only repos in the toolforge
organization for package building jobs.

Signed-off-by: David Caro <dcaro@wikimedia.org>
Change-Id: I01e1f50dd58beaad1e841284115d23b36441e336