run_tests: consider the tests passed even if there's skipped ones
It seems that the output of the failed tests might not be the last in
the string, and followed by a comma when skipping some, so we started
reporting them as failed even if they did not (on the deploy cookbook
side we are doing it ok though).

This matches now outputs like `0 failures, 5 skipped`.

Change-Id: Ib9c8c7aa3e73da0f02d33dc589dbd6bfc7b4068e
Signed-off-by: David Caro <dcaro@wikimedia.org>