wmcs_libs: k8s: Support tofu-managed groups for HAProxy
The K8s HAProxies require some security group rules that are not managed
by the cookbook, and the number of those will only increase as we move
responsibilities from the old front proxy servers to the HAProxies. Thus
I've sent a patch[0] to have the security group rules managed by
tofu-provisioning instead of continuing to manage them by hand (or
somehow adding security group rule management to the cookbooks).

[0]: https://gitlab.wikimedia.org/repos/cloud/toolforge/tofu-provisioning/-/merge_requests/84

The Tofu patch uses the 'service' Tofu module, which has a slightly
different naming convention for security and server groups than what
we're currently using for the K8s cluster. I've chosen to update the
cookbooks to use the Tofu-style names when provisioning new instances,
instead of hacking the Tofu code to support customizing the names to
what we're already using, for two reasons:
* long-term this will result in more consistent naming across k8s and
  non-k8s services in Toolforge, and
* hacking the Python code that generates these names is much easier than
  adding the new customization options in Tofu's own
  declarative-almost-programming-language.

Bug: T405078
Change-Id: I76492f775ae9898cd0540a8d64111cc1eca0ceb0