toolforge: k8s: Resolve K8s HAProxy VIPs from Hiera
For the upcoming migration to the dual-stack subnet, the VIP will be
different (and temporarily different nodes will have different VIPs, and
the DNS will point to the "wrong" one for new instances). So read the
correct VIP for a specific node from Hiera instead of relying on the
DNS name duplicated here.

Bug: T405078
Change-Id: Ieef9ba524afcf9bfa2ea1d3b4be625897b81aded