tofu: Allow running apply on a single cluster
Sometimes it's useful to run "--apply" on a single cluster, e.g. when
there is some issue in the codfw cluster but things are fine in the
eqiad cluster.

We already have the opentofu-infra-diff systemd timer that will alert us
if one of the two clusters is not in sync with the tofu-infra repo.

Bug: T411090
Change-Id: I091016321d392ef188ca6ec02d2aa6cf34e6e1e3